refactor: atomsをAtomsに変更
